+++ app/src/main/res/values/strings.xml
@@ -16,5 +16,4 @@
 -->
 <resources>
     <string name="app_name">Now in Android</string>
-    <string name="not_connected">⚠️ You aren’t connected to the internet</string>
 </resources>
