+++ feature/foryou/src/androidTest/java/com/google/samples/apps/nowinandroid/feature/foryou/ForYouScreenTest.kt
@@ -54,6 +54,7 @@ class ForYouScreenTest {
                     isSyncing = false,
                     onboardingUiState = OnboardingUiState.Loading,
                     feedState = NewsFeedUiState.Loading,
+                    isOffline = false,
                     onTopicCheckedChanged = { _, _ -> },
                     saveFollowedTopics = {},
                     onNewsResourcesCheckedChanged = { _, _ -> }
@@ -76,6 +77,7 @@ class ForYouScreenTest {
                     isSyncing = true,
                     onboardingUiState = OnboardingUiState.NotShown,
                     feedState = NewsFeedUiState.Success(emptyList()),
+                    isOffline = false,
                     onTopicCheckedChanged = { _, _ -> },
                     saveFollowedTopics = {},
                     onNewsResourcesCheckedChanged = { _, _ -> }
@@ -103,6 +105,7 @@ class ForYouScreenTest {
                     feedState = NewsFeedUiState.Success(
                         feed = emptyList()
                     ),
+                    isOffline = false,
                     onTopicCheckedChanged = { _, _ -> },
                     saveFollowedTopics = {},
                     onNewsResourcesCheckedChanged = { _, _ -> }
@@ -146,6 +149,7 @@ class ForYouScreenTest {
                     feedState = NewsFeedUiState.Success(
                         feed = emptyList()
                     ),
+                    isOffline = false,
                     onTopicCheckedChanged = { _, _ -> },
                     saveFollowedTopics = {},
                     onNewsResourcesCheckedChanged = { _, _ -> }
@@ -182,6 +186,7 @@ class ForYouScreenTest {
                     onboardingUiState =
                     OnboardingUiState.Shown(topics = testTopics),
                     feedState = NewsFeedUiState.Loading,
+                    isOffline = false,
                     onTopicCheckedChanged = { _, _ -> },
                     saveFollowedTopics = {},
                     onNewsResourcesCheckedChanged = { _, _ -> }
@@ -204,6 +209,7 @@ class ForYouScreenTest {
                     isSyncing = false,
                     onboardingUiState = OnboardingUiState.NotShown,
                     feedState = NewsFeedUiState.Loading,
+                    isOffline = false,
                     onTopicCheckedChanged = { _, _ -> },
                     saveFollowedTopics = {},
                     onNewsResourcesCheckedChanged = { _, _ -> }
@@ -229,6 +235,7 @@ class ForYouScreenTest {
                         SaveableNewsResource(it, false)
                     }
                 ),
+                isOffline = false,
                 onTopicCheckedChanged = { _, _ -> },
                 saveFollowedTopics = {},
                 onNewsResourcesCheckedChanged = { _, _ -> }
