+++ feature/foryou/src/main/java/com/google/samples/apps/nowinandroid/feature/foryou/ForYouScreen.kt
@@ -33,54 +33,35 @@ import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeDrawing
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.windowInsetsBottomHeight
 import androidx.compose.foundation.lazy.LazyListScope
-import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.GridCells.Adaptive
 import androidx.compose.foundation.lazy.grid.GridItemSpan
 import androidx.compose.foundation.lazy.grid.LazyGridScope
-import androidx.compose.foundation.lazy.grid.LazyHorizontalGrid
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.lazy.grid.rememberLazyGridState
-import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.layout.layout
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.max
-import androidx.compose.ui.unit.sp
-import androidx.compose.ui.util.trace
 import androidx.core.view.doOnPreDraw
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.ExperimentalLifecycleComposeApi
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import coil.compose.AsyncImage
 import com.google.samples.apps.nowinandroid.core.designsystem.component.NiaFilledButton
 import com.google.samples.apps.nowinandroid.core.designsystem.component.NiaOverlayLoadingWheel
-import com.google.samples.apps.nowinandroid.core.designsystem.component.NiaToggleButton
-import com.google.samples.apps.nowinandroid.core.designsystem.icon.NiaIcons
 import com.google.samples.apps.nowinandroid.core.designsystem.theme.NiaTheme
 import com.google.samples.apps.nowinandroid.core.domain.model.FollowableTopic
 import com.google.samples.apps.nowinandroid.core.domain.model.SaveableNewsResource
@@ -90,6 +71,10 @@ import com.google.samples.apps.nowinandroid.core.ui.DevicePreviews
 import com.google.samples.apps.nowinandroid.core.ui.NewsFeedUiState
 import com.google.samples.apps.nowinandroid.core.ui.TrackScrollJank
 import com.google.samples.apps.nowinandroid.core.ui.newsFeed
+import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState.LoadFailed
+import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState.Loading
+import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState.NotShown
+import com.google.samples.apps.nowinandroid.feature.foryou.components.TopicSelection
 
 @OptIn(ExperimentalLifecycleComposeApi::class)
 @Composable
@@ -100,30 +85,34 @@ internal fun ForYouRoute(
     val onboardingUiState by viewModel.onboardingUiState.collectAsStateWithLifecycle()
     val feedState by viewModel.feedState.collectAsStateWithLifecycle()
     val isSyncing by viewModel.isSyncing.collectAsStateWithLifecycle()
+    val isOffline by viewModel.isOffline.collectAsStateWithLifecycle()
 
     ForYouScreen(
+        modifier = modifier,
         isSyncing = isSyncing,
         onboardingUiState = onboardingUiState,
         feedState = feedState,
+        isOffline = isOffline,
         onTopicCheckedChanged = viewModel::updateTopicSelection,
         saveFollowedTopics = viewModel::dismissOnboarding,
-        onNewsResourcesCheckedChanged = viewModel::updateNewsResourceSaved,
-        modifier = modifier
+        onNewsResourcesCheckedChanged = viewModel::updateNewsResourceSaved
     )
 }
 
 @Composable
 internal fun ForYouScreen(
+    modifier: Modifier = Modifier,
     isSyncing: Boolean,
     onboardingUiState: OnboardingUiState,
     feedState: NewsFeedUiState,
+    isOffline: Boolean,
     onTopicCheckedChanged: (String, Boolean) -> Unit,
     saveFollowedTopics: () -> Unit,
-    onNewsResourcesCheckedChanged: (String, Boolean) -> Unit,
-    modifier: Modifier = Modifier,
+    onNewsResourcesCheckedChanged: (String, Boolean) -> Unit
 ) {
-    val isOnboardingLoading = onboardingUiState is OnboardingUiState.Loading
+    val isOnboardingLoading = onboardingUiState is Loading
     val isFeedLoading = feedState is NewsFeedUiState.Loading
+    val isOnboardingNotShown = onboardingUiState is NotShown
 
     // Workaround to call Activity.reportFullyDrawn from Jetpack Compose.
     // This code should be called when the UI is ready for use
@@ -156,39 +145,47 @@ internal fun ForYouScreen(
             .testTag("forYou:feed"),
         state = state
     ) {
-        onboarding(
-            onboardingUiState = onboardingUiState,
-            onTopicCheckedChanged = onTopicCheckedChanged,
-            saveFollowedTopics = saveFollowedTopics,
-            // Custom LayoutModifier to remove the enforced parent 16.dp contentPadding
-            // from the LazyVerticalGrid and enable edge-to-edge scrolling for this section
-            interestsItemModifier = Modifier.layout { measurable, constraints ->
-                val placeable = measurable.measure(
-                    constraints.copy(
-                        maxWidth = constraints.maxWidth + 32.dp.roundToPx()
+        if (!isOffline) {
+            onboarding(
+                onboardingUiState = onboardingUiState,
+                onTopicCheckedChanged = onTopicCheckedChanged,
+                saveFollowedTopics = saveFollowedTopics,
+                // Custom LayoutModifier to remove the enforced parent 16.dp contentPadding
+                // from the LazyVerticalGrid and enable edge-to-edge scrolling for this section
+                interestsItemModifier = Modifier.layout { measurable, constraints ->
+                    val placeable = measurable.measure(
+                        constraints.copy(
+                            maxWidth = constraints.maxWidth + 32.dp.roundToPx()
+                        )
                     )
-                )
-                layout(placeable.width, placeable.height) {
-                    placeable.place(0, 0)
+                    layout(placeable.width, placeable.height) {
+                        placeable.place(0, 0)
+                    }
                 }
-            }
-        )
+            )
 
-        newsFeed(
-            feedState = feedState,
-            onNewsResourcesCheckedChanged = onNewsResourcesCheckedChanged,
-        )
+            newsFeed(
+                feedState = feedState,
+                onNewsResourcesCheckedChanged = onNewsResourcesCheckedChanged,
+            )
 
-        item(span = { GridItemSpan(maxLineSpan) }) {
-            Column {
-                Spacer(modifier = Modifier.height(8.dp))
-                // Add space for the content to clear the "offline" snackbar.
-                // TODO: Check that the Scaffold handles this correctly in NiaApp
-                // if (isOffline) Spacer(modifier = Modifier.height(48.dp))
-                Spacer(Modifier.windowInsetsBottomHeight(WindowInsets.safeDrawing))
+            item(span = { GridItemSpan(maxLineSpan) }) {
+                Column {
+                    Spacer(modifier = Modifier.height(8.dp))
+                    // TODO: Check that the Scaffold handles this correctly in NiaApp
+                    Spacer(Modifier.windowInsetsBottomHeight(WindowInsets.safeDrawing))
+                }
             }
         }
     }
+
+    /**
+     * Show the Offline Screen if there is no internet connection.
+     */
+    if (isOffline) {
+        OfflineScreen()
+    }
+
     AnimatedVisibility(
         visible = isSyncing || isFeedLoading || isOnboardingLoading,
         enter = slideInVertically(
@@ -199,9 +196,7 @@ internal fun ForYouScreen(
         ) + fadeOut(),
     ) {
         val loadingContentDescription = stringResource(id = R.string.for_you_loading)
-        Box(
-            modifier = Modifier.fillMaxWidth()
-        ) {
+        Box(modifier = Modifier.fillMaxWidth()) {
             NiaOverlayLoadingWheel(
                 modifier = Modifier
                     .align(Alignment.Center)
@@ -222,169 +217,55 @@ private fun LazyGridScope.onboarding(
     onTopicCheckedChanged: (String, Boolean) -> Unit,
     saveFollowedTopics: () -> Unit,
     interestsItemModifier: Modifier = Modifier
-) {
-    when (onboardingUiState) {
-        OnboardingUiState.Loading,
-        OnboardingUiState.LoadFailed,
-        OnboardingUiState.NotShown -> Unit
+) = when (onboardingUiState) {
+    Loading,
+    LoadFailed,
+    NotShown -> Unit
 
-        is OnboardingUiState.Shown -> {
-            item(span = { GridItemSpan(maxLineSpan) }) {
-                Column(modifier = interestsItemModifier) {
-                    Text(
-                        text = stringResource(R.string.onboarding_guidance_title),
-                        textAlign = TextAlign.Center,
+    is OnboardingUiState.Shown -> {
+        item(span = { GridItemSpan(maxLineSpan) }) {
+            Column(modifier = interestsItemModifier) {
+                Text(
+                    text = stringResource(R.string.onboarding_guidance_title),
+                    textAlign = TextAlign.Center,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = 24.dp),
+                    style = MaterialTheme.typography.titleMedium
+                )
+                Text(
+                    text = stringResource(R.string.onboarding_guidance_subtitle),
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(top = 8.dp, start = 16.dp, end = 16.dp),
+                    textAlign = TextAlign.Center,
+                    style = MaterialTheme.typography.bodyMedium
+                )
+                TopicSelection(
+                    onboardingUiState,
+                    onTopicCheckedChanged,
+                    Modifier.padding(bottom = 8.dp)
+                )
+                // Done button
+                Row(
+                    horizontalArrangement = Arrangement.Center,
+                    modifier = Modifier.fillMaxWidth()
+                ) {
+                    NiaFilledButton(
+                        onClick = saveFollowedTopics,
+                        enabled = onboardingUiState.isDismissable,
                         modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 24.dp),
-                        style = MaterialTheme.typography.titleMedium
-                    )
-                    Text(
-                        text = stringResource(R.string.onboarding_guidance_subtitle),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(top = 8.dp, start = 16.dp, end = 16.dp),
-                        textAlign = TextAlign.Center,
-                        style = MaterialTheme.typography.bodyMedium
-                    )
-                    TopicSelection(
-                        onboardingUiState,
-                        onTopicCheckedChanged,
-                        Modifier.padding(bottom = 8.dp)
-                    )
-                    // Done button
-                    Row(
-                        horizontalArrangement = Arrangement.Center,
-                        modifier = Modifier.fillMaxWidth()
+                            .padding(horizontal = 40.dp)
+                            .width(364.dp)
                     ) {
-                        NiaFilledButton(
-                            onClick = saveFollowedTopics,
-                            enabled = onboardingUiState.isDismissable,
-                            modifier = Modifier
-                                .padding(horizontal = 40.dp)
-                                .width(364.dp)
-                        ) {
-                            Text(
-                                text = stringResource(R.string.done)
-                            )
-                        }
+                        Text(
+                            text = stringResource(R.string.done)
+                        )
                     }
                 }
             }
         }
     }
-}
-
-@Composable
-private fun TopicSelection(
-    onboardingUiState: OnboardingUiState.Shown,
-    onTopicCheckedChanged: (String, Boolean) -> Unit,
-    modifier: Modifier = Modifier
-) = trace("TopicSelection") {
-    val lazyGridState = rememberLazyGridState()
-    TrackScrollJank(scrollableState = lazyGridState, stateName = "forYou:TopicSelection")
-
-    LazyHorizontalGrid(
-        state = lazyGridState,
-        rows = GridCells.Fixed(3),
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp),
-        contentPadding = PaddingValues(24.dp),
-        modifier = modifier
-            // LazyHorizontalGrid has to be constrained in height.
-            // However, we can't set a fixed height because the horizontal grid contains
-            // vertical text that can be rescaled.
-            // When the fontScale is at most 1, we know that the horizontal grid will be at most
-            // 240dp tall, so this is an upper bound for when the font scale is at most 1.
-            // When the fontScale is greater than 1, the height required by the text inside the
-            // horizontal grid will increase by at most the same factor, so 240sp is a valid
-            // upper bound for how much space we need in that case.
-            // The maximum of these two bounds is therefore a valid upper bound in all cases.
-            .heightIn(max = max(240.dp, with(LocalDensity.current) { 240.sp.toDp() }))
-            .fillMaxWidth()
-    ) {
-        items(onboardingUiState.topics) {
-            SingleTopicButton(
-                name = it.topic.name,
-                topicId = it.topic.id,
-                imageUrl = it.topic.imageUrl,
-                isSelected = it.isFollowed,
-                onClick = onTopicCheckedChanged
-            )
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun SingleTopicButton(
-    name: String,
-    topicId: String,
-    imageUrl: String,
-    isSelected: Boolean,
-    onClick: (String, Boolean) -> Unit
-) = trace("SingleTopicButton") {
-    Surface(
-        modifier = Modifier
-            .width(312.dp)
-            .heightIn(min = 56.dp),
-        shape = RoundedCornerShape(corner = CornerSize(8.dp)),
-        color = MaterialTheme.colorScheme.surface,
-        selected = isSelected,
-        onClick = {
-            onClick(topicId, !isSelected)
-        }
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.padding(start = 12.dp, end = 8.dp)
-        ) {
-            TopicIcon(
-                imageUrl = imageUrl
-            )
-            Text(
-                text = name,
-                style = MaterialTheme.typography.titleSmall,
-                modifier = Modifier
-                    .padding(horizontal = 12.dp)
-                    .weight(1f),
-                color = MaterialTheme.colorScheme.onSurface
-            )
-            NiaToggleButton(
-                checked = isSelected,
-                onCheckedChange = { checked -> onClick(topicId, checked) },
-                icon = {
-                    Icon(
-                        imageVector = NiaIcons.Add,
-                        contentDescription = name
-                    )
-                },
-                checkedIcon = {
-                    Icon(
-                        imageVector = NiaIcons.Check,
-                        contentDescription = name
-                    )
-                }
-            )
-        }
-    }
-}
-
-@Composable
-fun TopicIcon(
-    imageUrl: String,
-    modifier: Modifier = Modifier
-) {
-    AsyncImage(
-        // TODO b/228077205, show loading image visual instead of static placeholder
-        placeholder = painterResource(R.drawable.ic_icon_placeholder),
-        model = imageUrl,
-        contentDescription = null, // decorative
-        colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.primary),
-        modifier = modifier
-            .padding(10.dp)
-            .size(32.dp)
-    )
 }
 
 @DevicePreviews
@@ -400,6 +281,7 @@ fun ForYouScreenPopulatedFeed() {
                         SaveableNewsResource(it, false)
                     }
                 ),
+                isOffline = false,
                 onTopicCheckedChanged = { _, _ -> },
                 saveFollowedTopics = {},
                 onNewsResourcesCheckedChanged = { _, _ -> }
@@ -421,6 +303,7 @@ fun ForYouScreenOfflinePopulatedFeed() {
                         SaveableNewsResource(it, false)
                     }
                 ),
+                isOffline = true,
                 onTopicCheckedChanged = { _, _ -> },
                 saveFollowedTopics = {},
                 onNewsResourcesCheckedChanged = { _, _ -> }
@@ -444,6 +327,7 @@ fun ForYouScreenTopicSelection() {
                         SaveableNewsResource(it, false)
                     }
                 ),
+                isOffline = false,
                 onTopicCheckedChanged = { _, _ -> },
                 saveFollowedTopics = {},
                 onNewsResourcesCheckedChanged = { _, _ -> }
@@ -459,8 +343,9 @@ fun ForYouScreenLoading() {
         NiaTheme {
             ForYouScreen(
                 isSyncing = false,
-                onboardingUiState = OnboardingUiState.Loading,
+                onboardingUiState = Loading,
                 feedState = NewsFeedUiState.Loading,
+                isOffline = false,
                 onTopicCheckedChanged = { _, _ -> },
                 saveFollowedTopics = {},
                 onNewsResourcesCheckedChanged = { _, _ -> }
@@ -476,12 +361,13 @@ fun ForYouScreenPopulatedAndLoading() {
         NiaTheme {
             ForYouScreen(
                 isSyncing = true,
-                onboardingUiState = OnboardingUiState.Loading,
+                onboardingUiState = Loading,
                 feedState = NewsFeedUiState.Success(
                     feed = previewNewsResources.map {
                         SaveableNewsResource(it, false)
                     }
                 ),
+                isOffline = false,
                 onTopicCheckedChanged = { _, _ -> },
                 saveFollowedTopics = {},
                 onNewsResourcesCheckedChanged = { _, _ -> }
