+++ feature/foryou/src/main/java/com/google/samples/apps/nowinandroid/feature/foryou/components/SingleTopicButton.kt
@@ -1,0 +1,91 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.feature.foryou.components
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CornerSize
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.trace
+import com.google.samples.apps.nowinandroid.core.designsystem.component.NiaToggleButton
+import com.google.samples.apps.nowinandroid.core.designsystem.icon.NiaIcons
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun SingleTopicButton(
+    name: String,
+    topicId: String,
+    imageUrl: String,
+    isSelected: Boolean,
+    onClick: (String, Boolean) -> Unit
+) = trace("SingleTopicButton") {
+    Surface(
+        modifier = Modifier
+            .width(312.dp)
+            .heightIn(min = 56.dp),
+        shape = RoundedCornerShape(corner = CornerSize(8.dp)),
+        color = MaterialTheme.colorScheme.surface,
+        selected = isSelected,
+        onClick = {
+            onClick(topicId, !isSelected)
+        }
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier.padding(start = 12.dp, end = 8.dp)
+        ) {
+            TopicIcon(
+                imageUrl = imageUrl
+            )
+            Text(
+                text = name,
+                style = MaterialTheme.typography.titleSmall,
+                modifier = Modifier
+                    .padding(horizontal = 12.dp)
+                    .weight(1f),
+                color = MaterialTheme.colorScheme.onSurface
+            )
+            NiaToggleButton(
+                checked = isSelected,
+                onCheckedChange = { checked -> onClick(topicId, checked) },
+                icon = {
+                    Icon(
+                        imageVector = NiaIcons.Add,
+                        contentDescription = name
+                    )
+                },
+                checkedIcon = {
+                    Icon(
+                        imageVector = NiaIcons.Check,
+                        contentDescription = name
+                    )
+                }
+            )
+        }
+    }
+}
