+++ feature/foryou/src/main/java/com/google/samples/apps/nowinandroid/feature/foryou/components/TopicSelection.kt
@@ -1,0 +1,75 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.feature.foryou.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.lazy.grid.GridCells.Fixed
+import androidx.compose.foundation.lazy.grid.LazyHorizontalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.max
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.util.trace
+import com.google.samples.apps.nowinandroid.core.ui.TrackScrollJank
+import com.google.samples.apps.nowinandroid.feature.foryou.OnboardingUiState
+
+@Composable
+fun TopicSelection(
+    onboardingUiState: OnboardingUiState.Shown,
+    onTopicCheckedChanged: (String, Boolean) -> Unit,
+    modifier: Modifier = Modifier
+) = trace("TopicSelection") {
+    val lazyGridState = rememberLazyGridState()
+    TrackScrollJank(scrollableState = lazyGridState, stateName = "forYou:TopicSelection")
+
+    LazyHorizontalGrid(
+        state = lazyGridState,
+        rows = Fixed(3),
+        horizontalArrangement = Arrangement.spacedBy(12.dp),
+        verticalArrangement = Arrangement.spacedBy(12.dp),
+        contentPadding = PaddingValues(24.dp),
+        modifier = modifier
+            // LazyHorizontalGrid has to be constrained in height.
+            // However, we can't set a fixed height because the horizontal grid contains
+            // vertical text that can be rescaled.
+            // When the fontScale is at most 1, we know that the horizontal grid will be at most
+            // 240dp tall, so this is an upper bound for when the font scale is at most 1.
+            // When the fontScale is greater than 1, the height required by the text inside the
+            // horizontal grid will increase by at most the same factor, so 240sp is a valid
+            // upper bound for how much space we need in that case.
+            // The maximum of these two bounds is therefore a valid upper bound in all cases.
+            .heightIn(max = max(240.dp, with(LocalDensity.current) { 240.sp.toDp() }))
+            .fillMaxWidth()
+    ) {
+        items(onboardingUiState.topics) {
+            SingleTopicButton(
+                name = it.topic.name,
+                topicId = it.topic.id,
+                imageUrl = it.topic.imageUrl,
+                isSelected = it.isFollowed,
+                onClick = onTopicCheckedChanged
+            )
+        }
+    }
+}
