+++ feature/foryou/src/main/res/drawable/ic_error_no_internet.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2022 The Android Open Source Project
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="192dp"
+    android:height="192dp"
+    android:viewportWidth="64"
+    android:viewportHeight="64">
+    <path
+        android:pathData="M32,32m-31,0a31,31 0,1 1,62 0a31,31 0,1 1,-62 0"
+        android:strokeColor="#8B418F"
+        android:strokeWidth="2" />
+    <path
+        android:fillColor="#8B418F"
+        android:pathData="M32,49m-4,0a4,4 0,1 1,8 0a4,4 0,1 1,-8 0"
+        android:strokeWidth="2" />
+    <path
+        android:fillColor="#8B418F"
+        android:pathData="M32,11L32,11A4,4 0,0 1,36 15L36,34A4,4 0,0 1,32 38L32,38A4,4 0,0 1,28 34L28,15A4,4 0,0 1,32 11z"
+        android:strokeWidth="2" />
+</vector>
